requirements.txt
asgiref==3.2.7
coverage==5.1
Django==3.0.6
django-redis==4.11.0
djangorestframework==3.11.0
pytz==2020.1
redis==3.5.1
sqlparse==0.3.1
